fix issue with navigation in new capture, fix issue with permission screen leading to the wrong screen

+++ app/src/main/java/com/edgetech/bbscout/components/location/LocationAwareActivity.kt
@@ -32,9 +32,9 @@ abstract class LocationAwareActivity : ComponentActivity() {
 
                     if (loc == null) {
 
-                        appLocation.postValue(AppLocation(location.accuracy.toInt(), location.latitude, location.longitude))
+                        appLocation.postValue(AppLocation(location.accuracy.toDouble(), location.latitude, location.longitude))
                         loc = AppLocation(
-                            location.accuracy.toInt(),
+                            location.accuracy.toDouble(),
                             location.latitude,
                             location.longitude
                         )
